% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{auto_ddsPLS}
\alias{auto_ddsPLS}
\title{Automatic}
\usage{
auto_ddsPLS(
  Xs,
  Y,
  lambdas = NULL,
  tau = 0.01,
  R_max = NA,
  lambdas_super = NULL,
  method = 2,
  NZV = 0.001,
  scaleMat = T,
  keep_model_per_block = T,
  plotVarB = T
)
}
\arguments{
\item{Xs}{list}

\item{Y}{matrix}

\item{tau}{rate of learning}

\item{NZV}{near zero variance}

\item{keep_model_per_block}{must be put to true}

\item{plotVarB}{logical}

\item{lambda_min}{coef}

\item{n_lambdas}{number of coeffs}
}
\value{

}
\description{
Automatic
}
