% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_PLS_Bootstrap.R
\name{bootstrap_pls_CT}
\alias{bootstrap_pls_CT}
\title{Title}
\usage{
bootstrap_pls_CT(X_init, Y_init, h = 1, lambdas = 0, lambda_prev = NA, u, v)
}
\arguments{
\item{X_init}{matrix of covariates}

\item{Y_init}{matrix of response}

\item{h}{number of components to build}

\item{lambdas}{vector}

\item{lambda_prev}{lambdas used to build the previous components}

\item{u}{X weights of the previous components}

\item{v}{Y weights of the previous components}
}
\value{
Internal object
}
\description{
Title
}
