% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootstrap_Rcpp}
\alias{bootstrap_Rcpp}
\title{C++ implementation of the bootstrap operations}
\usage{
bootstrap_Rcpp(
  U,
  V,
  X,
  Y,
  lambdas,
  lambda_prev,
  R,
  n_B,
  doBoot,
  n,
  p,
  q,
  N_lambdas,
  lambda0
)
}
\arguments{
\item{U}{The weights for X part.}

\item{V}{The weights for Y part.}

\item{X}{The matrix of X part.}

\item{Y}{The matrix of X part.}

\item{lambdas}{The to be tested values for lambda.}

\item{lambda_prev}{The previously selected values for lambda.}

\item{R}{The number of components to build.}

\item{n_B}{The number of bootstrap samples to generate and analyse.}

\item{doBoot}{Wheteher do bootstrap operations.}

\item{n}{The number of observations.}

\item{p}{The number of variables of X part.}

\item{q}{The number of variables of Y part.}

\item{N_lambdas}{The number of to be tested values for lambda.}

\item{lambda0}{The vector of lambda0}
}
\description{
Start the bootstrap operations.
Should not be used by user.
}
