% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddspls.R
\name{bootstrapWrap}
\alias{bootstrapWrap}
\title{C++ wrapper for bootstrap function}
\usage{
bootstrapWrap(
  U,
  V,
  X,
  Y,
  lambdas,
  lambda_prev,
  R,
  n_B,
  doBoot = TRUE,
  n,
  p,
  q,
  n_lambdas,
  lambda0.
)
}
\arguments{
\item{U}{matrix, weights X}

\item{V}{matrix, weights Y}

\item{X}{matrix}

\item{Y}{matrix}

\item{lambdas}{vector, the to be tested values for lambda}

\item{lambda_prev}{vector, the previous selected values for lambda}

\item{R}{integer, the desired number of components}

\item{n_B}{integer, the number of bootstrap samples required}

\item{doBoot}{boolean, whether or not perform bootstrap. Used to build the
final model (FALSE)}

\item{n}{integer, the number of observations}

\item{p}{integer, the number of covariates}

\item{q}{integer, the number of response variables}

\item{n_lambdas}{integer, the number of to be tested lambdas}

\item{lambda0.}{the vector of lambda0}
}
\value{
A list
}
\description{
The wrapper used to start the bootstrap commands. Not to be used by the user.
}
