% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsPLS2.R
\name{ddsPLS2}
\alias{ddsPLS2}
\title{Title}
\usage{
ddsPLS2(
  Xs,
  Y,
  lam,
  tau = 0.0975,
  method = 2,
  NZV = 0.001,
  Rs = NA,
  standardize = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{lam}{lam}

\item{tau}{tau}

\item{NZV}{NZV}

\item{Rs}{Rs}
}
\value{

}
\description{
Title
}
