% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ddsPLS_App}
\alias{ddsPLS_App}
\title{Applet to start ddsPLS}
\usage{
ddsPLS_App(...)
}
\arguments{
\item{...}{Same parameters as ddsPLS}
}
\value{
Mainly visual objects, also possible to save plots
}
\description{
Applet to start ddsPLS
}
