% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{do_boot}
\alias{do_boot}
\title{Title}
\usage{
do_boot(x0, y0, Y_init, lam = 0, deflatX = T)
}
\arguments{
\item{x0}{x0}

\item{y0}{y0}

\item{lam}{lam}

\item{Rs}{Rs}

\item{NCORES}{number of cores}
}
\value{

}
\description{
Title
}
