% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{est_l_min}
\alias{est_l_min}
\title{Title}
\usage{
est_l_min(
  Xs,
  Y,
  names_blocks = NULL,
  posLegend = "topright",
  x.intersp = 0.7,
  y.intersp = 0.7
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{posLegend}{posLegend}

\item{x.intersp}{x.intersp}

\item{y.intersp}{y.intersp}
}
\value{

}
\description{
Title
}
