% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{model_PLS}
\alias{model_PLS}
\title{Title}
\usage{
model_PLS(
  x,
  y,
  lam,
  deflatX = T,
  R = 1,
  remove_COV = NULL,
  RSS0 = NULL,
  NZV = 0.001,
  to.scale = T,
  COV_init = NULL
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{deflatX}{deflatX}

\item{R}{R}

\item{NZV}{NZV}

\item{to.scale}{to.scale}
}
\value{

}
\description{
Title
}
