% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{perf_auto_ddsPLS}
\alias{perf_auto_ddsPLS}
\title{Title}
\usage{
perf_auto_ddsPLS(
  Xs,
  Y,
  lambdas = 0.5,
  lambdas_super = NULL,
  tau = 0.01,
  kfolds = "loo",
  method = 2,
  fold_fixed = NULL,
  NCORES = 1,
  NZV = 0.001
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{tau}{rate learning}

\item{kfolds}{number of folds}

\item{fold_fixed}{yes or no}

\item{NCORES}{number of cores}

\item{NZV}{near zero var}

\item{n_rate}{number of rates}

\item{rate_min}{min rate}

\item{rate_max}{max rate}

\item{R_max}{max}

\item{lambda_min}{min lambda}

\item{n_lambdas}{number of lambdas}
}
\value{

}
\description{
Title
}
