% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mddsPLS.R
\name{plot.mddsPLS}
\alias{plot.mddsPLS}
\title{Function to plot \emph{mddsPLS}}
\usage{
\method{plot}{mddsPLS}(x, vizu = "weights", super = FALSE,
  mar_left = 2, block = NULL, comp = NULL, addY = FALSE,
  pos_legend = "topright", legend_names = NULL, ...)
}
\arguments{
\item{x}{The perf_mddsPLS object.}

\item{vizu}{character. One of \emph{weights}, \emph{heatmap}, \emph{components}, \emph{correlogram}}

\item{super}{logical. If \emph{TRUE} barplots are filled with **Super-Weights** in the case of \emph{vizu=weights} of with général **X** and **Y** components else.}

\item{mar_left}{positive float. Extra lines to add to the left margins, where the variable names are written.}

\item{block}{vector of intergers indicating which components must be plotted. If equals \emph{NULL} then all the components are plotted. Initialized to \emph{NULL}.}

\item{comp}{vector of intergers indicating which blocks must be plotted. If equals \emph{NULL} then all the blocks are plotted. Initialized to \emph{NULL}.}

\item{addY}{logical. Whether or not to plot **Block Y**. Initialized to \emph{FALSE}.}

\item{pos_legend}{Initialized to "topright"}

\item{legend_names}{vector of character. Indicates the names of the blocks. Initialized to NULL and in this case just gets positions in the Xs list.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
The plot visualisation
}
\description{
That function must be applied to a \emph{mddsPLS} object. Extra parameters are
 avalaible to control the plot quality.
}
\examples{
library(doParallel)
# Classification example :
data("penicilliumYES")
X <- penicilliumYES$X
X <- scale(X[,which(apply(X,2,sd)>0)])
Y <- as.factor(unlist(lapply(c("Melanoconidiu","Polonicum","Venetum"),
function(tt){rep(tt,12)})))
x <- mddsPLS(Xs = X,Y = Y,R = 3, mode = "clas",lambda=0.8)
plot(x)

# Regression example :
data("liver.toxicity")
X <- scale(liver.toxicity$gene)
Y <- scale(liver.toxicity$clinic)
#res_cv_reg <- ddsPLS(Xs = X,Y = Y,lambda=0.8,R = 2)
#plot(res_cv_reg)
}
\seealso{
\code{\link{mddsPLS}}, \code{\link{summary.mddsPLS}}
}
