% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sparse_PLS_Bootstrap.R
\name{plot.sparse_PLS_Bootstrap}
\alias{plot.sparse_PLS_Bootstrap}
\title{Function to plot bootstrap performance results of the ddsPLS algorithm.}
\usage{
\method{plot}{sparse_PLS_Bootstrap}(x, h_opt = NULL, type = "p", lty = NA, no_plot = F, ...)
}
\arguments{
\item{x}{A sparse_PLS_Bootstrap object}

\item{h_opt}{Optimal number of components to be plotted}

\item{type}{Type of the weights and regression coefficients}

\item{lty}{Shape of the lines, if type=="l"}

\item{no_plot}{Internal logical parameter}

\item{...}{Other plotting parameters to affect the plot.}
}
\description{
Function to plot bootstrap performance results of the ddsPLS algorithm.
}
