% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{plot_all_in_one}
\alias{plot_all_in_one}
\title{Title}
\usage{
plot_all_in_one(
  model,
  names_blocks = NULL,
  ymax_super = NULL,
  cexLegend = 1,
  pos = "bottom",
  singul = NULL
)
}
\arguments{
\item{model}{model}

\item{names_blocks}{names_blocks}

\item{ymax_super}{ymax_super}

\item{cexLegend}{cexLegend}

\item{pos}{pos}

\item{ymax}{ymax}
}
\value{

}
\description{
Title
}
