% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{plot_res_optim}
\alias{plot_res_optim}
\title{Title}
\usage{
plot_res_optim(res, colsK = "black", ymin = 0, ymax = 5, sig_f = NULL)
}
\arguments{
\item{res}{res}

\item{colsK}{colsK}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{sig_f}{sig_f}
}
\value{

}
\description{
Title
}
