% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_ddsPLS.R
\name{plot_var_B}
\alias{plot_var_B}
\title{Title}
\usage{
plot_var_B(
  model,
  posLegend = "bottomright",
  plotSuper = TRUE,
  x.intersp = 0.7,
  y.intersp = 0.7,
  names_blocks = NULL,
  low_lambda = NULL
)
}
\arguments{
\item{model}{model}

\item{posLegend}{position of the legend}

\item{x.intersp}{legend x space}

\item{y.intersp}{legend y space}
}
\value{

}
\description{
Title
}
