% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddsPLS2.R
\name{predict_ddsPLS2}
\alias{predict_ddsPLS2}
\title{Title}
\usage{
predict_ddsPLS2(model, xs)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{position of the legend}

\item{x.intersp}{legend x space}

\item{y.intersp}{legend y space}
}
\value{

}
\description{
Title
}
