% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ddsPLS.R
\name{print.ddsPLS}
\alias{print.ddsPLS}
\title{Function to sum up bootstrap performance results of the ddsPLS algorithm}
\usage{
\method{print}{ddsPLS}(x, ...)
}
\arguments{
\item{x}{A ddsPLS object.}

\item{...}{Other parameters to be taken into account.}
}
\description{
Function to sum up bootstrap performance results of the ddsPLS algorithm
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{plot.ddsPLS}}, \code{\link{predict.ddsPLS}}
}
