% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{Q2_boot_sPLS}
\alias{Q2_boot_sPLS}
\title{Title}
\usage{
Q2_boot_sPLS(
  Xs,
  Y,
  keepXs = 1,
  keepYs = 1,
  n_B = 20,
  whichCriterion = "Q2",
  deflatX = T,
  NCORES = 1,
  center = T,
  NZV = 1e-09,
  verbose = F,
  sparse = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{keepXs}{keepXs}

\item{n_B}{n_B}

\item{deflatX}{deflatX}

\item{NCORES}{number of cores}

\item{center}{center}

\item{NZV}{NZV}

\item{verbose}{verbose}

\item{sparse}{sparse}
}
\value{

}
\description{
Title
}
