% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{Q2_global_ddsPLS}
\alias{Q2_global_ddsPLS}
\title{Title}
\usage{
Q2_global_ddsPLS(
  Xs,
  Y,
  lambdas = 0.5,
  tau = 0.0975,
  NCORES = 1,
  NZV = 0.001,
  method = 2
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{lambdas}{lambdas}

\item{tau}{rate learning}

\item{NCORES}{number of cores}

\item{NZV}{near zero var}
}
\value{

}
\description{
Title
}
