% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{Q2_local_ddsPLS}
\alias{Q2_local_ddsPLS}
\title{Title}
\usage{
Q2_local_ddsPLS(
  Xs,
  Y,
  N_lambdas = 100,
  lambda_max = 1,
  use_lambda = T,
  deflatX = T,
  tau = 0.0975,
  NCORES = 1,
  center = T,
  NZV = 0.001,
  verbose = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{tau}{rate learning}

\item{NCORES}{number of cores}

\item{NZV}{near zero var}

\item{verbose}{verbose}

\item{lambdas}{lambdas}
}
\value{

}
\description{
Title
}
