% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{Q2_UNIK_ddsPLS}
\alias{Q2_UNIK_ddsPLS}
\title{Title}
\usage{
Q2_UNIK_ddsPLS(
  Xs,
  Y,
  lambdas = NULL,
  n_B = 20,
  lowExplainedVariance = 0,
  deflatX = T,
  NCORES = 1,
  center = T,
  verbose = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{lambdas}{lambdas}

\item{NCORES}{number of cores}

\item{verbose}{verbose}
}
\value{

}
\description{
Title
}
