% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_PLS_Bootstrap.R
\name{sparse_PLS_Bootstrap}
\alias{sparse_PLS_Bootstrap}
\title{Title}
\usage{
sparse_PLS_Bootstrap(
  Xs,
  Y,
  lambdas = NULL,
  n_B = 20,
  lowExplainedVariance = 0,
  deflatX = T,
  NCORES = 1,
  center = T,
  verbose = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{lambdas}{list}

\item{n_B}{NUmber of Bootstrap samples}

\item{lowExplainedVariance}{Q_2 value above which component is not build. Default to "0".}

\item{deflatX}{wether or not to deflat X. Default to TRUE. Do not change unless good reasons.}

\item{NCORES}{number of cores}

\item{center}{wether or not to center X and Y. Default to TRUE. Do not change unless good reasons.}

\item{verbose}{verbose}
}
\value{
A "sparse_PLS_Bootstrap" object with results of bootstrap simulations
}
\description{
Title
}
