% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ddsPLS.R
\name{summary.ddsPLS}
\alias{summary.ddsPLS}
\title{Function to sum up bootstrap performance results of the ddsPLS algorithm}
\usage{
\method{summary}{ddsPLS}(
  object,
  return = FALSE,
  plotSelection = FALSE,
  las = 1,
  cex.names = 1,
  digits = 2,
  ...
)
}
\arguments{
\item{object}{A ddsPLS object.}

\item{return}{Wether or not to return the printed values, default to FALSE.}

\item{plotSelection}{boolean. Whether plot the selection variables.}

\item{las}{interger. Parameter for angle of variable names.}

\item{cex.names}{real positive. Which factor zomm the variable names.}

\item{digits}{integer indicating the number of decimal places (round) to be used.}

\item{...}{Other parameters to be taken into account.}
}
\description{
Function to sum up bootstrap performance results of the ddsPLS algorithm
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{plot.ddsPLS}}, \code{\link{predict.ddsPLS}}
}
