% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mddsPLS.R
\name{summary.mddsPLS}
\alias{summary.mddsPLS}
\title{The summary method of the \emph{mddsPLS} function.}
\usage{
\method{summary}{mddsPLS}(object, plot_present_indiv = TRUE, ...)
}
\arguments{
\item{object}{The object of class mddsPLS}

\item{plot_present_indiv}{logical. If \emph{TRUE}, plots a venn diagram of the missing
individuals in each \emph{X} dataset. If the dataset does not appear, it means that it
 does not have missing values}

\item{...}{Other parameters.}
}
\description{
This function is easy to use and gives information about the dataset and the model.
}
\examples{
library(ddsPLS)
data("liver.toxicity")
X <- scale(liver.toxicity$gene)
Y <- scale(liver.toxicity$clinic)
X1<-X[,1:10];X1[1,]<-NA
X2<-X[,11:20];X2[2:5,]<-NA
X3<-X[,21:30];X3[4:20,]<-NA
X4<-X[,31:40]
Xs <- list(x1=X1,x2=X2,aaaa=X3,X4)
# object <- mddsPLS(Xs = Xs,Y = Y[,1],lambda=0.1,R = 1, mode = "reg",verbose = TRUE)
# summary(object,plot_present_indiv = T)
}
\seealso{
\code{\link{mddsPLS}}
}
