% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.perf_mddsPLS.R
\name{summary.perf_mddsPLS}
\alias{summary.perf_mddsPLS}
\title{The summary method of the \strong{perf_mddsPLS} function.}
\usage{
\method{summary}{perf_mddsPLS}(object, plot_res_cv = T, ...)
}
\arguments{
\item{object}{The object of class mddsPLS}

\item{plot_res_cv}{logical. If \strong{TRUE}, plots the results of the cross-validation}

\item{...}{Other parameters.}
}
\description{
This function is easy to use and gives information about the dataset and the model.
}
\examples{
library(ddsPLS)
data("liverToxicity")
X <- scale(liverToxicity$gene)
Y <- scale(liverToxicity$clinic)
X1<-X[,1:10];X1[1,]<-NA
X2<-X[,11:20];X2[2:5,]<-NA
X3<-X[,21:30];X3[4:20,]<-NA
X4<-X[,31:40]
Xs <- list(x1=X1,x2=X2,aaaa=X3,X4)
# object <- perf_mddsPLS(Xs = Xs,Y = Y[,1], lambdas=c(0.1,0.2,0.3),R = 1,
# mode = "reg",kfolds=5)
# summary(object)
}
\seealso{
\code{\link{perf_mddsPLS}}, \code{\link{plot.perf_mddsPLS}}
}
