% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_PLS_Bootstrap.R
\name{bootstrap_pls_CT}
\alias{bootstrap_pls_CT}
\title{Title}
\usage{
bootstrap_pls_CT(X_init, Y_init, h = 1, lambdas = 0, lambda_prev = NA, u, v)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{type}{"CT" and "l1}

\item{paras}{list}

\item{NCORES}{number of cores}

\item{verbose}{verbose}
}
\value{

}
\description{
Title
}
