% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ddsPLS_App}
\alias{ddsPLS_App}
\title{Applet to start ddsPLS}
\usage{
ddsPLS_App(...)
}
\arguments{
\item{...}{No parameter is needed explicitly for this app.}
}
\value{
Nothing due to the nature of the applet.
}
\description{
Applet to start ddsPLS
}
