% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{do_loo}
\alias{do_loo}
\title{Title}
\usage{
do_loo(xs0, y0, lam = 0, kX = NULL, NCORES = 1, deflatX = T)
}
\arguments{
\item{xs0}{xs0}

\item{y0}{y0}

\item{lam}{lam}

\item{NCORES}{number of cores}

\item{Rs}{Rs}
}
\value{

}
\description{
Title
}
