% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q2_cbind.R
\name{do_one_component}
\alias{do_one_component}
\title{Title}
\usage{
do_one_component(
  x0,
  y0,
  n,
  p,
  q,
  COV,
  abs_COV,
  max_COV,
  lam,
  remove_COV = NULL
)
}
\arguments{
\item{x0}{x0}

\item{y0}{y0}

\item{n}{n}

\item{p}{p}

\item{q}{q}

\item{COV}{COV}

\item{abs_COV}{abs_COV}

\item{max_COV}{max_COV}

\item{lam}{lam}
}
\value{

}
\description{
Title
}
