% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_PLS_Bootstrap.R
\name{model_PLS}
\alias{model_PLS}
\title{Title}
\usage{
model_PLS(x, y, lam, deflatX = T, R = 1, to.scale = T)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{lam}{Lambda coefficient}

\item{deflatX}{deflatX}

\item{R}{R}

\item{to.scale}{to.scale}
}
\value{
Internal object
}
\description{
Title
}
