% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ddsPLS.R
\name{plot.ddsPLS}
\alias{plot.ddsPLS}
\title{Function to plot bootstrap performance results of the ddsPLS algorithm}
\usage{
\method{plot}{ddsPLS}(
  x,
  type = "criterion",
  digits = 1,
  legend.position = "topright",
  horiz = TRUE,
  biPlot = FALSE,
  las = 0,
  col = NULL,
  cex.names = 1,
  mar = c(5, 4, 4, 2) + 0.1,
  ...
)
}
\arguments{
\item{x}{A ddsPLS object}

\item{type}{The type of graphics. One of "criterion" (default), "total",
"prop", "predict", "Q2r", "Q2", "R2r", "R2", "weightX", "weightY",
 "loadingX" or "loadingY".}

\item{digits}{double. Rounding of the written explained variance.}

\item{legend.position}{character. Where to put the legend.}

\item{horiz}{boolean. Whether to plot horizontally.}

\item{biPlot}{boolean wether or not to plot one component versus the other.}

\item{las}{numeric in (0,1,2,3): the style of axis labels.}

\item{col}{vector. Mainly to modify bars in weight plots.}

\item{cex.names}{double. Size factor for variable names.}

\item{mar}{vector. The margins for the plot.}

\item{...}{Other plotting parameters to affect the plot.}
}
\description{
Function to plot bootstrap performance results of the ddsPLS algorithm
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{predict.ddsPLS}}, \code{\link{summary.ddsPLS}}
}
