% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ddsPLS.R
\name{predict.ddsPLS}
\alias{predict.ddsPLS}
\title{Function to predict from ddsPLS objects}
\usage{
\method{predict}{ddsPLS}(
  object,
  X_test = NULL,
  toPlot = FALSE,
  doDiagnosis = T,
  legend.position = "topright",
  cex = 1,
  cex.text = 1,
  ...
)
}
\arguments{
\item{object}{A ddsPLS object.}

\item{X_test}{matrix, a test data-set. If is "NULL", the default value,
the predicted values for the train test are returned.}

\item{toPlot}{boolean, wether or not to plot the extreme value test plot.
Default to `TRUE`.}

\item{doDiagnosis}{Yes or no to perform diagnosis.}

\item{legend.position}{character. Where to put the legend.}

\item{cex}{float positive. Number indicating the amount by which plotting
symbols should be scaled relative to the default.}

\item{cex.text}{float positive. Number indicating the amount by which
plotting text elements should be scaled relative to the default.}

\item{...}{Other parameters}
}
\description{
Function to predict from ddsPLS objects
}
\seealso{
\code{\link{ddsPLS}}, \code{\link{plot.ddsPLS}}, \code{\link{summary.ddsPLS}}
}
