% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mddsPLS.R
\name{predict.mddsPLS}
\alias{predict.mddsPLS}
\title{The predict function of a mdd-sPLS model}
\usage{
\method{predict}{mddsPLS}(mod_0, newX)
}
\arguments{
\item{mod_0}{A mdd-sPLS object, output from the mddsPLS function.}

\item{newX}{A data-set where individuals are described by the same as for mod_0}
}
\value{
Predicted values for those individuals
}
\description{
The predict function of a mdd-sPLS model
}
\examples{
mod_0 <- mddsPLS(X,Y)
Y_est <- predict(mod_0,X)
}
