% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mddsPLS.R
\name{predict.mddsPLS}
\alias{predict.mddsPLS}
\title{The predict function of a mdd-sPLS model}
\usage{
\method{predict}{mddsPLS}(mod_0, newX)
}
\arguments{
\item{mod_0}{A mdd-sPLS object, output from the mddsPLS function.}

\item{newX}{A data-set where individuals are described by the same as for mod_0}
}
\value{
A matrix of estimated \emph{Y_test} values.
}
\description{
The predict function of a mdd-sPLS model
}
\examples{
## Not run:
data("liver.toxicity")
X <- scale(liver.toxicity$gene)
Y <- scale(liver.toxicity$clinic)
mod_0 <- mddsPLS(X,Y)
Y_test <- predict(mod_0,X)
## End(**Not run**)
}
