% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_PLS_Bootstrap.R
\name{sparse_PLS_Bootstrap}
\alias{sparse_PLS_Bootstrap}
\title{Title}
\usage{
sparse_PLS_Bootstrap(
  Xs,
  Y,
  type = "CT",
  paras = NULL,
  n_B = 20,
  lowExplainedVariance = 0,
  deflatX = T,
  NCORES = 1,
  center = T,
  verbose = F
)
}
\arguments{
\item{Xs}{Xs}

\item{Y}{Y}

\item{type}{"CT" and "l1}

\item{paras}{list}

\item{NCORES}{number of cores}

\item{verbose}{verbose}
}
\value{

}
\description{
Title
}
