% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mddsPLS.R
\name{summary.mddsPLS}
\alias{summary.mddsPLS}
\title{The summary method of the \strong{mddsPLS} function.}
\usage{
\method{summary}{mddsPLS}(object, main_plot_indiv = NULL,
  fontsize = 10, alpha = 0.7, ...)
}
\arguments{
\item{object}{The object of class mddsPLS}

\item{main_plot_indiv}{character. Main of the Venn diagram. Initialized to NULL.}

\item{fontsize}{interger. The size of the text, initialized to 10.}

\item{alpha}{real between 0 and 1. The transparency parameter.}

\item{...}{Other parameters.}
}
\description{
This function is easy to use and gives information about the dataset and the model.
}
\examples{
library(ddsPLS)
data("liverToxicity")
X <- scale(liverToxicity$gene)
Y <- scale(liverToxicity$clinic)
X1<-X[,1:10];X1[1,]<-NA
X2<-X[,11:20];X2[2:5,]<-NA
X3<-X[,21:30];X3[4:20,]<-NA
X4<-X[,31:40]
Xs <- list(x1=X1,x2=X2,aaaa=X3,X4)
# object <- mddsPLS(Xs = Xs,Y = Y[,1],lambda=0.1,R = 1, mode = "reg",verbose = TRUE)
# summary(object)
}
\seealso{
\code{\link{mddsPLS}}
}
